refactor: move ttf font loading into separate class
